+++ app/controllers/tools_controller.rb
@@ -100,12 +100,12 @@ class ToolsController < ApplicationController
   end
 
   def session_info
-    session[:show_session_info] = params[:show] == 'show'
+    session[:show_session_info] = (params[:show] == 'show')
     render :nothing => true
   end
   
   def groups_menu
-    session[:expand_group_menu] = params[:folding] == 'expand'
+    session[:expand_group_menu] = (params[:folding] == 'expand')
     render :nothing => true
   end