+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and other major modifications are listed here in order to give a short overview
 of every version. For earlier versions and more detail, please consult the
 commit history on GitHub.
 
+## Version X
+
+### Internal
+
+* added lockr.js for localStorage
+
 ## Version 2.0.0
 
 ### User